Add or modify Makefile recipes to enable regeneration of *.rdf files and retain expected *.ttl files content
`rdfpipe` utility is now used to regenerate *.rdf files what results in the
duplicated imports being discarded. Changes were made in Makefile file and
affects generation of *.rdf artefacts (owl-core, owl-restrictions and shacl
Makefile recipes) as well as conversion of these to TTL format
(convert-rdf-to-turtle). The latter were needed to keep the same TTL output
when using the new *.rdf files.

Scope of the changes:
* New generic convert-between-serialization-formats Makefile recipe
* The existing owl-core, owl-restrictions and shacl Makefile recipes now include an RDF regeneration step
* The existing convert-rdf-to-turtle Makefile recipe now uses the same namespaces that are used for XSLT transformation (test/ePO-default-config/namespaces.xml) to provide rdfpipe with prefixes so the output makes use of compact URIs.
* As the reported issue occurs for EPO v4.2.0 which introduces some new prefixes, these prefixes have also been added to test/ePO-default-config/namespaces.xml so it would be possible to reproduce the fix.