Implement generation of enriched namespaces file, reorganize project to depend on the new file
The new XML contains user namespaces (defined in namespaces.xml) and internal
namespaces (such as core-shape). Both model2owl internals and get_namespaces.sh
relies on that file and not on namespaces.xml. Moreover, model2owl internals now
depends on a new template which generates namespaces declarations based on
enriched file which is passed to the XSLT transformation engine as command line
argument. For now, the path of the new XML file is kept in the Makefile, but
this can be further improved in the future (if we come up with application
internal XSL config or other solution).

Scope of changes:
* new XSL file for generation of the enriched namespaces XML file
* unit test for the above XSL file
* project codebase:
  * new function generating namespaces declarations and references to
    `$namespacePrefixes` replaced with `$internalNamespacePrefixes`
* Makefile changes:
  * new recipe
  * existing 3 recipes generating *.rdf artefacts now include generation of the
    new XML file
  * required adjustments
* Adjustments in the get_namespaces.sh script