Add JSON-LD context generation module
Scope of changes:
* Implemented a new module for generating JSON-LD contexts from UML elements and selected connectors
* Added comprehensive unit tests for the new functionality
* Introduced a new Makefile recipe for JSON-LD context generation
* Added configuration parameter `generateReusedConceptsJSONLDcontext`
* Developed utility functions to streamline processing of relations in connectors (see `fetchers.xsl`)
* Added several helper utility functions
* Updated test data with minor improvements
* Updated README

Note: A new relation perspective (f:getRelationsFromConnector) was introduced to
simplify the processing of relations (object properties) encoded within
connectors. While UML permits multiple connectors with the same role for a
particular source class, the JSON-LD context allows only a single occurrence to
be defined. As a result, a more sophisticated processing logic was necessary.