Make the respec package generation logic generic and extendable
The package generation logic no longer assumes default values for
locations of model2owl project resources (SDS) and their labels
to be shown in the ReSpec HTML document. This is now expected to be
controlled by the user and therefore the corresponding properties
have been moved to metadata.json file.